Changes Books

+++ DAL/Entity/Book.cs
@@ -20,8 +20,7 @@ namespace DAL.Entity
         [Key]
         public int Id { get; set; }
         public int Isbn { get; set; }
-        public int Price { get; set; }
-
+        public double Price { get; set; }
         [StringLength(maximumLength:100)]
         public string Name { get; set; }
